Bug 1508863 - Initial work related to the New Tab Settings rework.
